Initialize signer lazily
By initializing the signer lazily, only when used the first time, we
also postpone the check for missing `secret_key` and `principal_id`.

This makes it possible to load the Rails environment without setting the
Azure Storage credentials in environments that do not need to connect to
Azure. For example, a build step that precompiles assets needs to load
the environments, but usually does not need Azure Storage. Before this
change, one would still have to set some (possibly fake) `secret_key` or
`principal_id` in such build step, while with this change those can be
omitted in any situation in which no real calls to the Azure API will be
made.